Struct aws_sdk_apigatewayv2::client::fluent_builders::UpdateAuthorizer

pub struct UpdateAuthorizer<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> UpdateAuthorizer<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<UpdateAuthorizerOutput, SdkError<UpdateAuthorizerError>> where
R::Policy: SmithyRetryPolicy<UpdateAuthorizerInputOperationOutputAlias, UpdateAuthorizerOutput, UpdateAuthorizerError, UpdateAuthorizerInputOperationRetryAlias>,

pub fn api_id(self, inp: impl Into<String>) -> Self

The API identifier.

pub fn set_api_id(self, input: Option<String>) -> Self

pub fn authorizer_credentials_arn(self, inp: impl Into<String>) -> Self

Specifies the required credentials as an IAM role for API Gateway to invoke the authorizer. To specify an IAM role for API Gateway to assume, use the role's Amazon Resource Name (ARN). To use resource-based permissions on the Lambda function, don't specify this parameter.

pub fn set_authorizer_credentials_arn(self, input: Option<String>) -> Self

pub fn authorizer_id(self, inp: impl Into<String>) -> Self

The authorizer identifier.

pub fn set_authorizer_id(self, input: Option<String>) -> Self

pub fn authorizer_payload_format_version(self, inp: impl Into<String>) -> Self

Specifies the format of the payload sent to an HTTP API Lambda authorizer. Required for HTTP API Lambda authorizers. Supported values are 1.0 and 2.0. To learn more, see Working with AWS Lambda authorizers for HTTP APIs.

pub fn set_authorizer_payload_format_version(
self,
input: Option<String>
) -> Self

pub fn authorizer_result_ttl_in_seconds(self, inp: i32) -> Self

The time to live (TTL) for cached authorizer results, in seconds. If it equals 0, authorization caching is disabled. If it is greater than 0, API Gateway caches authorizer responses. The maximum value is 3600, or 1 hour. Supported only for HTTP API Lambda authorizers.

pub fn set_authorizer_result_ttl_in_seconds(self, input: Option<i32>) -> Self

pub fn authorizer_type(self, inp: AuthorizerType) -> Self

The authorizer type. Specify REQUEST for a Lambda function using incoming request parameters. Specify JWT to use JSON Web Tokens (supported only for HTTP APIs).

pub fn set_authorizer_type(self, input: Option<AuthorizerType>) -> Self

pub fn authorizer_uri(self, inp: impl Into<String>) -> Self

The authorizer's Uniform Resource Identifier (URI). For REQUEST authorizers, this must be a well-formed Lambda function URI, for example, arn:aws:apigateway:us-west-2:lambda:path/2015-03-31/functions/arn:aws:lambda:us-west-2:{account_id}:function:{lambda_function_name}/invocations. In general, the URI has this form: arn:aws:apigateway:{region}:lambda:path/{service_api} , where {region} is the same as the region hosting the Lambda function, path indicates that the remaining substring in the URI should be treated as the path to the resource, including the initial /. For Lambda functions, this is usually of the form /2015-03-31/functions/[FunctionARN]/invocations. Supported only for REQUEST authorizers.

pub fn set_authorizer_uri(self, input: Option<String>) -> Self

pub fn enable_simple_responses(self, inp: bool) -> Self

Specifies whether a Lambda authorizer returns a response in a simple format. By default, a Lambda authorizer must return an IAM policy. If enabled, the Lambda authorizer can return a boolean value instead of an IAM policy. Supported only for HTTP APIs. To learn more, see Working with AWS Lambda authorizers for HTTP APIs

pub fn set_enable_simple_responses(self, input: Option<bool>) -> Self

pub fn identity_source(self, inp: impl Into<String>) -> Self

Appends an item to IdentitySource.

To override the contents of this collection use set_identity_source.

The identity source for which authorization is requested.

For a REQUEST authorizer, this is optional. The value is a set of one or more mapping expressions of the specified request parameters. The identity source can be headers, query string parameters, stage variables, and context parameters. For example, if an Auth header and a Name query string parameter are defined as identity sources, this value is route.request.header.Auth, route.request.querystring.Name for WebSocket APIs. For HTTP APIs, use selection expressions prefixed with $, for example, $request.header.Auth, $request.querystring.Name. These parameters are used to perform runtime validation for Lambda-based authorizers by verifying all of the identity-related request parameters are present in the request, not null, and non-empty. Only when this is true does the authorizer invoke the authorizer Lambda function. Otherwise, it returns a 401 Unauthorized response without calling the Lambda function. For HTTP APIs, identity sources are also used as the cache key when caching is enabled. To learn more, see Working with AWS Lambda authorizers for HTTP APIs.

For JWT, a single entry that specifies where to extract the JSON Web Token (JWT) from inbound requests. Currently only header-based and query parameter-based selections are supported, for example $request.header.Authorization.

pub fn set_identity_source(self, input: Option<Vec<String>>) -> Self

pub fn identity_validation_expression(self, inp: impl Into<String>) -> Self

This parameter is not used.

pub fn set_identity_validation_expression(self, input: Option<String>) -> Self

pub fn jwt_configuration(self, inp: JwtConfiguration) -> Self

Represents the configuration of a JWT authorizer. Required for the JWT authorizer type. Supported only for HTTP APIs.

pub fn set_jwt_configuration(self, input: Option<JwtConfiguration>) -> Self

pub fn name(self, inp: impl Into<String>) -> Self

The name of the authorizer.

pub fn set_name(self, input: Option<String>) -> Self