Struct aws_sdk_apigatewayv2::client::fluent_builders::UpdateRoute

pub struct UpdateRoute<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> UpdateRoute<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(self) -> Result<UpdateRouteOutput, SdkError<UpdateRouteError>> where
R::Policy: SmithyRetryPolicy<UpdateRouteInputOperationOutputAlias, UpdateRouteOutput, UpdateRouteError, UpdateRouteInputOperationRetryAlias>,

pub fn api_id(self, inp: impl Into<String>) -> Self

The API identifier.

pub fn set_api_id(self, input: Option<String>) -> Self

pub fn api_key_required(self, inp: bool) -> Self

Specifies whether an API key is required for the route. Supported only for WebSocket APIs.

pub fn set_api_key_required(self, input: Option<bool>) -> Self

pub fn authorization_scopes(self, inp: impl Into<String>) -> Self

Appends an item to AuthorizationScopes.

To override the contents of this collection use set_authorization_scopes.

The authorization scopes supported by this route.

pub fn set_authorization_scopes(self, input: Option<Vec<String>>) -> Self

pub fn authorization_type(self, inp: AuthorizationType) -> Self

The authorization type for the route. For WebSocket APIs, valid values are NONE for open access, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer For HTTP APIs, valid values are NONE for open access, JWT for using JSON Web Tokens, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer.

pub fn set_authorization_type(self, input: Option<AuthorizationType>) -> Self

pub fn authorizer_id(self, inp: impl Into<String>) -> Self

The identifier of the Authorizer resource to be associated with this route. The authorizer identifier is generated by API Gateway when you created the authorizer.

pub fn set_authorizer_id(self, input: Option<String>) -> Self

pub fn model_selection_expression(self, inp: impl Into<String>) -> Self

The model selection expression for the route. Supported only for WebSocket APIs.

pub fn set_model_selection_expression(self, input: Option<String>) -> Self

pub fn operation_name(self, inp: impl Into<String>) -> Self

The operation name for the route.

pub fn set_operation_name(self, input: Option<String>) -> Self

pub fn request_models(self, k: impl Into<String>, v: impl Into<String>) -> Self

Adds a key-value pair to RequestModels.

To override the contents of this collection use set_request_models.

The request models for the route. Supported only for WebSocket APIs.

pub fn set_request_models(self, input: Option<HashMap<String, String>>) -> Self

pub fn request_parameters(
self,
k: impl Into<String>,
v: impl Into<ParameterConstraints>
) -> Self

Adds a key-value pair to RequestParameters.

To override the contents of this collection use set_request_parameters.

The request parameters for the route. Supported only for WebSocket APIs.

pub fn set_request_parameters(
self,
input: Option<HashMap<String, ParameterConstraints>>
) -> Self

pub fn route_id(self, inp: impl Into<String>) -> Self

The route ID.

pub fn set_route_id(self, input: Option<String>) -> Self

pub fn route_key(self, inp: impl Into<String>) -> Self

The route key for the route.

pub fn set_route_key(self, input: Option<String>) -> Self

pub fn route_response_selection_expression(self, inp: impl Into<String>) -> Self

The route response selection expression for the route. Supported only for WebSocket APIs.

pub fn set_route_response_selection_expression(
self,
input: Option<String>
) -> Self

pub fn target(self, inp: impl Into<String>) -> Self

The target for the route.

pub fn set_target(self, input: Option<String>) -> Self