Struct aws_sdk_apigatewayv2::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for ApiGatewayV2.

This client allows ergonomic access to a ApiGatewayV2-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn create_api(&self) -> CreateApi<C, M, R>

Constructs a fluent builder for the CreateApi operation.

See CreateApi for more information about the operation and its arguments.

pub fn create_api_mapping(&self) -> CreateApiMapping<C, M, R>

Constructs a fluent builder for the CreateApiMapping operation.

See CreateApiMapping for more information about the operation and its arguments.

pub fn create_authorizer(&self) -> CreateAuthorizer<C, M, R>

Constructs a fluent builder for the CreateAuthorizer operation.

See CreateAuthorizer for more information about the operation and its arguments.

pub fn create_deployment(&self) -> CreateDeployment<C, M, R>

Constructs a fluent builder for the CreateDeployment operation.

See CreateDeployment for more information about the operation and its arguments.

pub fn create_domain_name(&self) -> CreateDomainName<C, M, R>

Constructs a fluent builder for the CreateDomainName operation.

See CreateDomainName for more information about the operation and its arguments.

pub fn create_integration(&self) -> CreateIntegration<C, M, R>

Constructs a fluent builder for the CreateIntegration operation.

See CreateIntegration for more information about the operation and its arguments.

pub fn create_integration_response(&self) -> CreateIntegrationResponse<C, M, R>

Constructs a fluent builder for the CreateIntegrationResponse operation.

See CreateIntegrationResponse for more information about the operation and its arguments.

pub fn create_model(&self) -> CreateModel<C, M, R>

Constructs a fluent builder for the CreateModel operation.

See CreateModel for more information about the operation and its arguments.

pub fn create_route(&self) -> CreateRoute<C, M, R>

Constructs a fluent builder for the CreateRoute operation.

See CreateRoute for more information about the operation and its arguments.

pub fn create_route_response(&self) -> CreateRouteResponse<C, M, R>

Constructs a fluent builder for the CreateRouteResponse operation.

See CreateRouteResponse for more information about the operation and its arguments.

pub fn create_stage(&self) -> CreateStage<C, M, R>

Constructs a fluent builder for the CreateStage operation.

See CreateStage for more information about the operation and its arguments.

pub fn create_vpc_link(&self) -> CreateVpcLink<C, M, R>

Constructs a fluent builder for the CreateVpcLink operation.

See CreateVpcLink for more information about the operation and its arguments.

pub fn delete_access_log_settings(&self) -> DeleteAccessLogSettings<C, M, R>

Constructs a fluent builder for the DeleteAccessLogSettings operation.

See DeleteAccessLogSettings for more information about the operation and its arguments.

pub fn delete_api(&self) -> DeleteApi<C, M, R>

Constructs a fluent builder for the DeleteApi operation.

See DeleteApi for more information about the operation and its arguments.

pub fn delete_api_mapping(&self) -> DeleteApiMapping<C, M, R>

Constructs a fluent builder for the DeleteApiMapping operation.

See DeleteApiMapping for more information about the operation and its arguments.

pub fn delete_authorizer(&self) -> DeleteAuthorizer<C, M, R>

Constructs a fluent builder for the DeleteAuthorizer operation.

See DeleteAuthorizer for more information about the operation and its arguments.

pub fn delete_cors_configuration(&self) -> DeleteCorsConfiguration<C, M, R>

Constructs a fluent builder for the DeleteCorsConfiguration operation.

See DeleteCorsConfiguration for more information about the operation and its arguments.

pub fn delete_deployment(&self) -> DeleteDeployment<C, M, R>

Constructs a fluent builder for the DeleteDeployment operation.

See DeleteDeployment for more information about the operation and its arguments.

pub fn delete_domain_name(&self) -> DeleteDomainName<C, M, R>

Constructs a fluent builder for the DeleteDomainName operation.

See DeleteDomainName for more information about the operation and its arguments.

pub fn delete_integration(&self) -> DeleteIntegration<C, M, R>

Constructs a fluent builder for the DeleteIntegration operation.

See DeleteIntegration for more information about the operation and its arguments.

pub fn delete_integration_response(&self) -> DeleteIntegrationResponse<C, M, R>

Constructs a fluent builder for the DeleteIntegrationResponse operation.

See DeleteIntegrationResponse for more information about the operation and its arguments.

pub fn delete_model(&self) -> DeleteModel<C, M, R>

Constructs a fluent builder for the DeleteModel operation.

See DeleteModel for more information about the operation and its arguments.

pub fn delete_route(&self) -> DeleteRoute<C, M, R>

Constructs a fluent builder for the DeleteRoute operation.

See DeleteRoute for more information about the operation and its arguments.

pub fn delete_route_request_parameter(
&self
) -> DeleteRouteRequestParameter<C, M, R>

Constructs a fluent builder for the DeleteRouteRequestParameter operation.

See DeleteRouteRequestParameter for more information about the operation and its arguments.

pub fn delete_route_response(&self) -> DeleteRouteResponse<C, M, R>

Constructs a fluent builder for the DeleteRouteResponse operation.

See DeleteRouteResponse for more information about the operation and its arguments.

pub fn delete_route_settings(&self) -> DeleteRouteSettings<C, M, R>

Constructs a fluent builder for the DeleteRouteSettings operation.

See DeleteRouteSettings for more information about the operation and its arguments.

pub fn delete_stage(&self) -> DeleteStage<C, M, R>

Constructs a fluent builder for the DeleteStage operation.

See DeleteStage for more information about the operation and its arguments.

pub fn delete_vpc_link(&self) -> DeleteVpcLink<C, M, R>

Constructs a fluent builder for the DeleteVpcLink operation.

See DeleteVpcLink for more information about the operation and its arguments.

pub fn export_api(&self) -> ExportApi<C, M, R>

Constructs a fluent builder for the ExportApi operation.

See ExportApi for more information about the operation and its arguments.

pub fn get_api(&self) -> GetApi<C, M, R>

Constructs a fluent builder for the GetApi operation.

See GetApi for more information about the operation and its arguments.

pub fn get_api_mapping(&self) -> GetApiMapping<C, M, R>

Constructs a fluent builder for the GetApiMapping operation.

See GetApiMapping for more information about the operation and its arguments.

pub fn get_api_mappings(&self) -> GetApiMappings<C, M, R>

Constructs a fluent builder for the GetApiMappings operation.

See GetApiMappings for more information about the operation and its arguments.

pub fn get_apis(&self) -> GetApis<C, M, R>

Constructs a fluent builder for the GetApis operation.

See GetApis for more information about the operation and its arguments.

pub fn get_authorizer(&self) -> GetAuthorizer<C, M, R>

Constructs a fluent builder for the GetAuthorizer operation.

See GetAuthorizer for more information about the operation and its arguments.

pub fn get_authorizers(&self) -> GetAuthorizers<C, M, R>

Constructs a fluent builder for the GetAuthorizers operation.

See GetAuthorizers for more information about the operation and its arguments.

pub fn get_deployment(&self) -> GetDeployment<C, M, R>

Constructs a fluent builder for the GetDeployment operation.

See GetDeployment for more information about the operation and its arguments.

pub fn get_deployments(&self) -> GetDeployments<C, M, R>

Constructs a fluent builder for the GetDeployments operation.

See GetDeployments for more information about the operation and its arguments.

pub fn get_domain_name(&self) -> GetDomainName<C, M, R>

Constructs a fluent builder for the GetDomainName operation.

See GetDomainName for more information about the operation and its arguments.

pub fn get_domain_names(&self) -> GetDomainNames<C, M, R>

Constructs a fluent builder for the GetDomainNames operation.

See GetDomainNames for more information about the operation and its arguments.

pub fn get_integration(&self) -> GetIntegration<C, M, R>

Constructs a fluent builder for the GetIntegration operation.

See GetIntegration for more information about the operation and its arguments.

pub fn get_integration_response(&self) -> GetIntegrationResponse<C, M, R>

Constructs a fluent builder for the GetIntegrationResponse operation.

See GetIntegrationResponse for more information about the operation and its arguments.

pub fn get_integration_responses(&self) -> GetIntegrationResponses<C, M, R>

Constructs a fluent builder for the GetIntegrationResponses operation.

See GetIntegrationResponses for more information about the operation and its arguments.

pub fn get_integrations(&self) -> GetIntegrations<C, M, R>

Constructs a fluent builder for the GetIntegrations operation.

See GetIntegrations for more information about the operation and its arguments.

pub fn get_model(&self) -> GetModel<C, M, R>

Constructs a fluent builder for the GetModel operation.

See GetModel for more information about the operation and its arguments.

pub fn get_models(&self) -> GetModels<C, M, R>

Constructs a fluent builder for the GetModels operation.

See GetModels for more information about the operation and its arguments.

pub fn get_model_template(&self) -> GetModelTemplate<C, M, R>

Constructs a fluent builder for the GetModelTemplate operation.

See GetModelTemplate for more information about the operation and its arguments.

pub fn get_route(&self) -> GetRoute<C, M, R>

Constructs a fluent builder for the GetRoute operation.

See GetRoute for more information about the operation and its arguments.

pub fn get_route_response(&self) -> GetRouteResponse<C, M, R>

Constructs a fluent builder for the GetRouteResponse operation.

See GetRouteResponse for more information about the operation and its arguments.

pub fn get_route_responses(&self) -> GetRouteResponses<C, M, R>

Constructs a fluent builder for the GetRouteResponses operation.

See GetRouteResponses for more information about the operation and its arguments.

pub fn get_routes(&self) -> GetRoutes<C, M, R>

Constructs a fluent builder for the GetRoutes operation.

See GetRoutes for more information about the operation and its arguments.

pub fn get_stage(&self) -> GetStage<C, M, R>

Constructs a fluent builder for the GetStage operation.

See GetStage for more information about the operation and its arguments.

pub fn get_stages(&self) -> GetStages<C, M, R>

Constructs a fluent builder for the GetStages operation.

See GetStages for more information about the operation and its arguments.

pub fn get_tags(&self) -> GetTags<C, M, R>

Constructs a fluent builder for the GetTags operation.

See GetTags for more information about the operation and its arguments.

pub fn get_vpc_link(&self) -> GetVpcLink<C, M, R>

Constructs a fluent builder for the GetVpcLink operation.

See GetVpcLink for more information about the operation and its arguments.

pub fn get_vpc_links(&self) -> GetVpcLinks<C, M, R>

Constructs a fluent builder for the GetVpcLinks operation.

See GetVpcLinks for more information about the operation and its arguments.

pub fn import_api(&self) -> ImportApi<C, M, R>

Constructs a fluent builder for the ImportApi operation.

See ImportApi for more information about the operation and its arguments.

pub fn reimport_api(&self) -> ReimportApi<C, M, R>

Constructs a fluent builder for the ReimportApi operation.

See ReimportApi for more information about the operation and its arguments.

pub fn reset_authorizers_cache(&self) -> ResetAuthorizersCache<C, M, R>

Constructs a fluent builder for the ResetAuthorizersCache operation.

See ResetAuthorizersCache for more information about the operation and its arguments.

pub fn tag_resource(&self) -> TagResource<C, M, R>

Constructs a fluent builder for the TagResource operation.

See TagResource for more information about the operation and its arguments.

pub fn untag_resource(&self) -> UntagResource<C, M, R>

Constructs a fluent builder for the UntagResource operation.

See UntagResource for more information about the operation and its arguments.

pub fn update_api(&self) -> UpdateApi<C, M, R>

Constructs a fluent builder for the UpdateApi operation.

See UpdateApi for more information about the operation and its arguments.

pub fn update_api_mapping(&self) -> UpdateApiMapping<C, M, R>

Constructs a fluent builder for the UpdateApiMapping operation.

See UpdateApiMapping for more information about the operation and its arguments.

pub fn update_authorizer(&self) -> UpdateAuthorizer<C, M, R>

Constructs a fluent builder for the UpdateAuthorizer operation.

See UpdateAuthorizer for more information about the operation and its arguments.

pub fn update_deployment(&self) -> UpdateDeployment<C, M, R>

Constructs a fluent builder for the UpdateDeployment operation.

See UpdateDeployment for more information about the operation and its arguments.

pub fn update_domain_name(&self) -> UpdateDomainName<C, M, R>

Constructs a fluent builder for the UpdateDomainName operation.

See UpdateDomainName for more information about the operation and its arguments.

pub fn update_integration(&self) -> UpdateIntegration<C, M, R>

Constructs a fluent builder for the UpdateIntegration operation.

See UpdateIntegration for more information about the operation and its arguments.

pub fn update_integration_response(&self) -> UpdateIntegrationResponse<C, M, R>

Constructs a fluent builder for the UpdateIntegrationResponse operation.

See UpdateIntegrationResponse for more information about the operation and its arguments.

pub fn update_model(&self) -> UpdateModel<C, M, R>

Constructs a fluent builder for the UpdateModel operation.

See UpdateModel for more information about the operation and its arguments.

pub fn update_route(&self) -> UpdateRoute<C, M, R>

Constructs a fluent builder for the UpdateRoute operation.

See UpdateRoute for more information about the operation and its arguments.

pub fn update_route_response(&self) -> UpdateRouteResponse<C, M, R>

Constructs a fluent builder for the UpdateRouteResponse operation.

See UpdateRouteResponse for more information about the operation and its arguments.

pub fn update_stage(&self) -> UpdateStage<C, M, R>

Constructs a fluent builder for the UpdateStage operation.

See UpdateStage for more information about the operation and its arguments.

pub fn update_vpc_link(&self) -> UpdateVpcLink<C, M, R>

Constructs a fluent builder for the UpdateVpcLink operation.

See UpdateVpcLink for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.