Module aws_sdk_apigatewayv2::error

Modules

access_denied_exception

See AccessDeniedException

bad_request_exception

See BadRequestException

conflict_exception

See ConflictException

not_found_exception

See NotFoundException

too_many_requests_exception

See TooManyRequestsException

Structs

AccessDeniedException

BadRequestException

The request is not valid, for example, the input is incomplete or incorrect. See the accompanying error message for details.

ConflictException

The requested operation would cause a conflict with the current state of a service resource associated with the request. Resolve the conflict before retrying this request. See the accompanying error message for details.

CreateApiError

CreateApiMappingError

CreateAuthorizerError

CreateDeploymentError

CreateDomainNameError

CreateIntegrationError

CreateIntegrationResponseError

CreateModelError

CreateRouteError

CreateRouteResponseError

CreateStageError

CreateVpcLinkError

DeleteAccessLogSettingsError

DeleteApiError

DeleteApiMappingError

DeleteAuthorizerError

DeleteCorsConfigurationError

DeleteDeploymentError

DeleteDomainNameError

DeleteIntegrationError

DeleteIntegrationResponseError

DeleteModelError

DeleteRouteError

DeleteRouteRequestParameterError

DeleteRouteResponseError

DeleteRouteSettingsError

DeleteStageError

DeleteVpcLinkError

ExportApiError

GetApiError

GetApiMappingError

GetApiMappingsError

GetApisError

GetAuthorizerError

GetAuthorizersError

GetDeploymentError

GetDeploymentsError

GetDomainNameError

GetDomainNamesError

GetIntegrationError

GetIntegrationResponseError

GetIntegrationResponsesError

GetIntegrationsError

GetModelError

GetModelTemplateError

GetModelsError

GetRouteError

GetRouteResponseError

GetRouteResponsesError

GetRoutesError

GetStageError

GetStagesError

GetTagsError

GetVpcLinkError

GetVpcLinksError

ImportApiError

NotFoundException

The resource specified in the request was not found. See the message field for more information.

ReimportApiError

ResetAuthorizersCacheError

TagResourceError

TooManyRequestsException

A limit has been exceeded. See the accompanying error message for details.

UntagResourceError

UpdateApiError

UpdateApiMappingError

UpdateAuthorizerError

UpdateDeploymentError

UpdateDomainNameError

UpdateIntegrationError

UpdateIntegrationResponseError

UpdateModelError

UpdateRouteError

UpdateRouteResponseError

UpdateStageError

UpdateVpcLinkError

Enums

CreateApiErrorKind

CreateApiMappingErrorKind

CreateAuthorizerErrorKind

CreateDeploymentErrorKind

CreateDomainNameErrorKind

CreateIntegrationErrorKind

CreateIntegrationResponseErrorKind

CreateModelErrorKind

CreateRouteErrorKind

CreateRouteResponseErrorKind

CreateStageErrorKind

CreateVpcLinkErrorKind

DeleteAccessLogSettingsErrorKind

DeleteApiErrorKind

DeleteApiMappingErrorKind

DeleteAuthorizerErrorKind

DeleteCorsConfigurationErrorKind

DeleteDeploymentErrorKind

DeleteDomainNameErrorKind

DeleteIntegrationErrorKind

DeleteIntegrationResponseErrorKind

DeleteModelErrorKind

DeleteRouteErrorKind

DeleteRouteRequestParameterErrorKind

DeleteRouteResponseErrorKind

DeleteRouteSettingsErrorKind

DeleteStageErrorKind

DeleteVpcLinkErrorKind

ExportApiErrorKind

GetApiErrorKind

GetApiMappingErrorKind

GetApiMappingsErrorKind

GetApisErrorKind

GetAuthorizerErrorKind

GetAuthorizersErrorKind

GetDeploymentErrorKind

GetDeploymentsErrorKind

GetDomainNameErrorKind

GetDomainNamesErrorKind

GetIntegrationErrorKind

GetIntegrationResponseErrorKind

GetIntegrationResponsesErrorKind

GetIntegrationsErrorKind

GetModelErrorKind

GetModelTemplateErrorKind

GetModelsErrorKind

GetRouteErrorKind

GetRouteResponseErrorKind

GetRouteResponsesErrorKind

GetRoutesErrorKind

GetStageErrorKind

GetStagesErrorKind

GetTagsErrorKind

GetVpcLinkErrorKind

GetVpcLinksErrorKind

ImportApiErrorKind

ReimportApiErrorKind

ResetAuthorizersCacheErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateApiErrorKind

UpdateApiMappingErrorKind

UpdateAuthorizerErrorKind

UpdateDeploymentErrorKind

UpdateDomainNameErrorKind

UpdateIntegrationErrorKind

UpdateIntegrationResponseErrorKind

UpdateModelErrorKind

UpdateRouteErrorKind

UpdateRouteResponseErrorKind

UpdateStageErrorKind

UpdateVpcLinkErrorKind