Module aws_sdk_apigatewayv2::model

Modules

access_log_settings

See AccessLogSettings

api

See Api

api_mapping

See ApiMapping

authorizer

See Authorizer

cors

See Cors

deployment

See Deployment

domain_name

See DomainName

domain_name_configuration

See DomainNameConfiguration

integration

See Integration

integration_response

See IntegrationResponse

jwt_configuration

See JwtConfiguration

model

See Model

mutual_tls_authentication

See MutualTlsAuthentication

mutual_tls_authentication_input

See MutualTlsAuthenticationInput

parameter_constraints

See ParameterConstraints

route

See Route

route_response

See RouteResponse

route_settings

See RouteSettings

stage

See Stage

tls_config

See TlsConfig

tls_config_input

See TlsConfigInput

vpc_link

See VpcLink

Structs

AccessLogSettings

Settings for logging access in a stage.

Api

Represents an API.

ApiMapping

Represents an API mapping.

Authorizer

Represents an authorizer.

Cors

Represents a CORS configuration. Supported only for HTTP APIs. See Configuring CORS for more information.

Deployment

An immutable representation of an API that can be called by users. A Deployment must be associated with a Stage for it to be callable over the internet.

DomainName

Represents a domain name.

DomainNameConfiguration

The domain name configuration.

Integration

Represents an integration.

IntegrationResponse

Represents an integration response.

JwtConfiguration

Represents the configuration of a JWT authorizer. Required for the JWT authorizer type. Supported only for HTTP APIs.

Model

Represents a data model for an API. Supported only for WebSocket APIs. See Create Models and Mapping Templates for Request and Response Mappings.

MutualTlsAuthentication

MutualTlsAuthenticationInput

ParameterConstraints

Validation constraints imposed on parameters of a request (path, query string, headers).

Route

Represents a route.

RouteResponse

Represents a route response.

RouteSettings

Represents a collection of route settings.

Stage

Represents an API stage.

TlsConfig

The TLS configuration for a private integration. If you specify a TLS configuration, private integration traffic uses the HTTPS protocol. Supported only for HTTP APIs.

TlsConfigInput

The TLS configuration for a private integration. If you specify a TLS configuration, private integration traffic uses the HTTPS protocol. Supported only for HTTP APIs.

VpcLink

Represents a VPC link.

Enums

AuthorizationType

The authorization type. For WebSocket APIs, valid values are NONE for open access, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer. For HTTP APIs, valid values are NONE for open access, JWT for using JSON Web Tokens, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer.

AuthorizerType

The authorizer type. Specify REQUEST for a Lambda function using incoming request parameters. Specify JWT to use JSON Web Tokens (supported only for HTTP APIs).

ConnectionType

Represents a connection type.

ContentHandlingStrategy

Specifies how to handle response payload content type conversions. Supported only for WebSocket APIs.

DeploymentStatus

Represents a deployment status.

DomainNameStatus

The status of the domain name migration. The valid values are AVAILABLE, UPDATING, PENDING_CERTIFICATE_REIMPORT, and PENDING_OWNERSHIP_VERIFICATION. If the status is UPDATING, the domain cannot be modified further until the existing operation is complete. If it is AVAILABLE, the domain can be updated.

EndpointType

Represents an endpoint type.

IntegrationType

Represents an API method integration type.

LoggingLevel

The logging level.

PassthroughBehavior

Represents passthrough behavior for an integration response. Supported only for WebSocket APIs.

ProtocolType

Represents a protocol type.

SecurityPolicy

The Transport Layer Security (TLS) version of the security policy for this domain name. The valid values are TLS_1_0 and TLS_1_2.

VpcLinkStatus

The status of the VPC link.

VpcLinkVersion

The version of the VPC link.