Struct aws_sdk_apigatewayv2::model::integration::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Integration

Implementations

impl Builder

pub fn api_gateway_managed(self, input: bool) -> Self

Specifies whether an integration is managed by API Gateway. If you created an API using using quick create, the resulting integration is managed by API Gateway. You can update a managed integration, but you can't delete it.

pub fn set_api_gateway_managed(self, input: Option<bool>) -> Self

pub fn connection_id(self, input: impl Into<String>) -> Self

The ID of the VPC link for a private integration. Supported only for HTTP APIs.

pub fn set_connection_id(self, input: Option<String>) -> Self

pub fn connection_type(self, input: ConnectionType) -> Self

The type of the network connection to the integration endpoint. Specify INTERNET for connections through the public routable internet or VPC_LINK for private connections between API Gateway and resources in a VPC. The default value is INTERNET.

pub fn set_connection_type(self, input: Option<ConnectionType>) -> Self

pub fn content_handling_strategy(self, input: ContentHandlingStrategy) -> Self

Supported only for WebSocket APIs. Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

pub fn set_content_handling_strategy(
self,
input: Option<ContentHandlingStrategy>
) -> Self

pub fn credentials_arn(self, input: impl Into<String>) -> Self

Specifies the credentials required for the integration, if any. For AWS integrations, three options are available. To specify an IAM Role for API Gateway to assume, use the role's Amazon Resource Name (ARN). To require that the caller's identity be passed through from the request, specify the string arn:aws:iam::*:user/*. To use resource-based permissions on supported AWS services, specify null.

pub fn set_credentials_arn(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

Represents the description of an integration.

pub fn set_description(self, input: Option<String>) -> Self

pub fn integration_id(self, input: impl Into<String>) -> Self

Represents the identifier of an integration.

pub fn set_integration_id(self, input: Option<String>) -> Self

pub fn integration_method(self, input: impl Into<String>) -> Self

Specifies the integration's HTTP method type.

pub fn set_integration_method(self, input: Option<String>) -> Self

pub fn integration_response_selection_expression(
self,
input: impl Into<String>
) -> Self

The integration response selection expression for the integration. Supported only for WebSocket APIs. See Integration Response Selection Expressions.

pub fn set_integration_response_selection_expression(
self,
input: Option<String>
) -> Self

pub fn integration_subtype(self, input: impl Into<String>) -> Self

Supported only for HTTP API AWS_PROXY integrations. Specifies the AWS service action to invoke. To learn more, see Integration subtype reference.

pub fn set_integration_subtype(self, input: Option<String>) -> Self

pub fn integration_type(self, input: IntegrationType) -> Self

The integration type of an integration. One of the following:

AWS: for integrating the route or method request with an AWS service action, including the Lambda function-invoking action. With the Lambda function-invoking action, this is referred to as the Lambda custom integration. With any other AWS service action, this is known as AWS integration. Supported only for WebSocket APIs.

AWS_PROXY: for integrating the route or method request with a Lambda function or other AWS service action. This integration is also referred to as a Lambda proxy integration.

HTTP: for integrating the route or method request with an HTTP endpoint. This integration is also referred to as the HTTP custom integration. Supported only for WebSocket APIs.

HTTP_PROXY: for integrating the route or method request with an HTTP endpoint, with the client request passed through as-is. This is also referred to as HTTP proxy integration.

MOCK: for integrating the route or method request with API Gateway as a "loopback" endpoint without invoking any backend. Supported only for WebSocket APIs.

pub fn set_integration_type(self, input: Option<IntegrationType>) -> Self

pub fn integration_uri(self, input: impl Into<String>) -> Self

For a Lambda integration, specify the URI of a Lambda function.

For an HTTP integration, specify a fully-qualified URL.

For an HTTP API private integration, specify the ARN of an Application Load Balancer listener, Network Load Balancer listener, or AWS Cloud Map service. If you specify the ARN of an AWS Cloud Map service, API Gateway uses DiscoverInstances to identify resources. You can use query parameters to target specific resources. To learn more, see DiscoverInstances. For private integrations, all resources must be owned by the same AWS account.

pub fn set_integration_uri(self, input: Option<String>) -> Self

pub fn passthrough_behavior(self, input: PassthroughBehavior) -> Self

Specifies the pass-through behavior for incoming requests based on the Content-Type header in the request, and the available mapping templates specified as the requestTemplates property on the Integration resource. There are three valid values: WHEN_NO_MATCH, WHEN_NO_TEMPLATES, and NEVER. Supported only for WebSocket APIs.

WHEN_NO_MATCH passes the request body for unmapped content types through to the integration backend without transformation.

NEVER rejects unmapped content types with an HTTP 415 Unsupported Media Type response.

WHEN_NO_TEMPLATES allows pass-through when the integration has no content types mapped to templates. However, if there is at least one content type defined, unmapped content types will be rejected with the same HTTP 415 Unsupported Media Type response.

pub fn set_passthrough_behavior(
self,
input: Option<PassthroughBehavior>
) -> Self

pub fn payload_format_version(self, input: impl Into<String>) -> Self

Specifies the format of the payload sent to an integration. Required for HTTP APIs.

pub fn set_payload_format_version(self, input: Option<String>) -> Self

pub fn request_parameters(
self,
k: impl Into<String>,
v: impl Into<String>
) -> Self

pub fn set_request_parameters(
self,
input: Option<HashMap<String, String>>
) -> Self

pub fn request_templates(
self,
k: impl Into<String>,
v: impl Into<String>
) -> Self

pub fn set_request_templates(
self,
input: Option<HashMap<String, String>>
) -> Self

pub fn response_parameters(
self,
k: impl Into<String>,
v: impl Into<HashMap<String, String>>
) -> Self

pub fn set_response_parameters(
self,
input: Option<HashMap<String, HashMap<String, String>>>
) -> Self

pub fn template_selection_expression(self, input: impl Into<String>) -> Self

The template selection expression for the integration. Supported only for WebSocket APIs.

pub fn set_template_selection_expression(self, input: Option<String>) -> Self

pub fn timeout_in_millis(self, input: i32) -> Self

Custom timeout between 50 and 29,000 milliseconds for WebSocket APIs and between 50 and 30,000 milliseconds for HTTP APIs. The default timeout is 29 seconds for WebSocket APIs and 30 seconds for HTTP APIs.

pub fn set_timeout_in_millis(self, input: Option<i32>) -> Self

pub fn tls_config(self, input: TlsConfig) -> Self

The TLS configuration for a private integration. If you specify a TLS configuration, private integration traffic uses the HTTPS protocol. Supported only for HTTP APIs.

pub fn set_tls_config(self, input: Option<TlsConfig>) -> Self

pub fn build(self) -> Integration

Consumes the builder and constructs a Integration